[test] Add dependency to the embedded-stdlib
When not building `all`, like the `build-script` does in CI, the tests
did not have a dependency on `embedded-stdlib`, and those tests will
fail to pass if one was using CMake/Ninja to execute the tests after
something like `ninja clean`.

The dependency is added if the embedded stdlib is being built.