[Sema] Treat package decls as exported
Currently, a package symbol imported via @_implementationOnly
is permitted in a package interface declaration, and can be
accessed by its client, potentially leading to a ub. It should
not be allowed as the package symbol in such scenario should
be treated as internal details.

This PR addresses the issue by treating a package decl as exported,
as it is "public" within a package boundary.

Ref rdar://117586046