Incorporate documentation from the Swift book.
These macros used to be special literals, documented in "The Swift
Programming Language".  Now that they're special macros instead, the
documentation has moved into the standard library.

This content is what's already being published for these symbols -- the
DocC sidecar for the stdlib (outside this repo) currently includes the
content that this commit is moving here.  That sidecar currently
overrides what's in /// comments in source.