[Macros] Cache `SourceLocationConverter` in `ExportedSourceFile`
We need a `SourceLocationConverter` every time we create a `PluginMessage.Syntax` to know the source location of that syntax node within the source file. This means that we needed to re-build the line table of the entire source file multiple times for every macro that we expand. Cache it to improve performance.

rdar://119047550

# Conflicts:
#	lib/ASTGen/Sources/ASTGen/SourceFile.swift