[OnoneSimplify] Fixed isDestroyed(after:)
The previous implementation just checked that a value's only uses
besides the begin_borrow were destroys.  That's insufficient to say the
value is destroyed _after_ the borrow (i.e. that all its destroys are
dominated by the borrow).  Add the relevant dominance check.

It would be better not to commpute the dominance tree here, so that
performance issue can be fixed separately.

rdar://119873930