[Debug] Accept _debugDescription in DebugDescriptionMacro (#70728)
Some data types cannot modify their `description` or `debugDescription` properties, as 
modification to those properties could result in breakage to their users. To support 
these conditions, `DebugDescriptionMacro` should support and prioritize an independent 
property (one that that is not being reused).

This change adds support for `_debugDescription`. The macro will now prioritize the 
description properties in this order:

1. `_debugDescription`
2. `debugDescription`
3. `debugDescription`

rdar://120498021