Rely on the next/nextElement call in an async for..each for throws effects
Our decision procedure for what type of error is thrown during
iteration of an async sequence was effectively duplicated between the
checking for the next/nextElement call (which is synthesized) and
separate logic for the async for..in loop. The latter is strictly
redundant, so switch to relying only on the former.

This is staging for customizing the next/nextElement call further.