[Distributed] Destroy decoded parameters after executeDistributedTarget
We recently fixed a similar issue where the objects created from
decodeNextArgument in a distributed thunk were not properly destroyed.
The fix was in https://github.com/apple/swift/pull/65952

In that fix, we missed handling of Indirect_In_Guaranteed which also
need to be destroyed, this patch corrects that omission.

Resolves https://github.com/apple/swift/issues/70004
Resolves rdar://119943672
Resolves rdar://119446012