Sema: Fix logic error in TypeReprCycleCheckWalker handling of 'Self.Foo'
We don't want to throw out a witness if it has any mention of 'Self.Foo';
we must also check that 'Foo' is one of the associated types being
inferred. Otherwise, it may be a protocol type alias, in which case
we want to proceed with the witness.

This fixes a regression from a recent change, a576984e2f7910760.

Fixes rdar://problem/120743365.