Sema: Diagnose retroactive conditional conformances
Passing in the declared interface type to checkConformance() here
masked a silent failure where if the type declaration was generic
and the conformance conditional, the conditional requirement check
would fail. As a result, we did not diagnose the absence of
@retroactive, nor the unnecessary presence of it.

Since we only care about the existence of some conformance, we
can use lookupConformance() instead.