[Concurrency] [shims] Don't declare `exit` in  the concurrency shims
Don't delete the OS declaration of `exit` because the concurrency shimes aren't always imported, and so the shim declaration might not always be available.
Don't override the OS declaration of `exit` in the concurrency shims since we can't just delete the OS one. Instead, set up internal shims just for building Concurrency that forward declares `exit`.