NCGenerics: simplify `isInterfaceTypeNoncopyable`
We can greatly simplify our noncopyable checks if we skip running these
diagnostic checks while in SIL mode, where it's not really needed. The
challenge with doing the checks in SIL mode is that there is sometimes
no generic environment when doing these queries. Whether that itself is
simply a bug or fundamental limitation, I haven't investigated.