NCGenerics: isMoveOnly and reference ownership
ReferenceStorageType can only contain a type that is class-like, like a
SILBox or a class-constrained existential, etc.

While classes do not and are not planned to officially support being
noncopyable, we do still have a lingering experiemental feature that
allows classes to be marked `@_moveOnly`. Thus, the best way to handle
queries about a `ReferenceStorageType` being move-only or noncopyable is
 to check whether the referent is.