[Frontend][AST] Add min-runtime-version option.
**This is intended for use within the compiler build itself; it is not
envisaged that third parties will use it.**

The goal here is to avoid a situation where a single PR that adds code
to call a new runtime function makes the compiler hard to build on Linux
and/or Windows, simply because you need a Swift build installed that has
the new runtime function in order to perform a successful build.

On Darwin we already have a way around this, namely the OS version based
availability mechanism.  This PR adds an equivalent for Linux and Windows
based on the Swift version number itself; when building the Swift code
within the compiler proper, this will be set to match the version of the
compiler that's being used, and thereby provide a way to gate the use of
newer runtime APIs by the compiler during the build.

rdar://121522431